backend/requirements.txt
flask
opencv-python-headless
pillow
numpy
onnxruntime
ultralytics
requests
python-dotenv
google-generativeai
flask-cors

torch
torchvision
waitress

fastapi
modal-client

python-multipart>=0.0.9
modal-client>=0.57.0